factor out slider dir and utils
